# Conflicts:
#	composeApp/build.gradle.kts
#	composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/composables/FileImagePicker.kt
#	composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/dialogs/tabs/BackgroundSettingsTab.kt
#	composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/dialogs/tabs/BibleSettingsTab.kt
#	composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/dialogs/tabs/LowerThirdSettingsTab.kt
#	composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/dialogs/tabs/SongSettingsTab.kt
#	composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/tabs/AnnouncementsTab.kt
#	composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/tabs/MediaTab.kt
#	composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/tabs/PicturesTab.kt
#	composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/tabs/PresentationTab.kt
#	composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/tabs/ScheduleTab.kt
#	composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/utils/Constants.kt
#	composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/viewmodel/PicturesViewModel.kt
#	composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/viewmodel/ScheduleViewModel.kt

+++ composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/composables/ColorPickerField.kt
@@ -43,9 +43,9 @@ fun ColorPickerField(
                     // Get the parent window to ensure dialog appears on top
                     val parentWindow = Window.getWindows().firstOrNull { it.isActive }
                     val initialColor = java.awt.Color(
-                        currentColor.red.toInt(),
-                        currentColor.green.toInt(),
-                        currentColor.blue.toInt()
+                        currentColor.red,
+                        currentColor.green,
+                        currentColor.blue
                     )
                     val selectedColor = JColorChooser.showDialog(
                         parentWindow,
